requirements.txt
SQLAlchemy==2.0.21
alembic==1.12.0
fastapi==0.103.2
psycopg2==2.9.9
pydantic==2.4.2
pydantic-core==2.10.1
uvicorn==0.23.2
starlette~=0.27.0
python-dotenv~=1.0.0